+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaAny.kt
@@ -1,0 +1,30 @@
+package br.com.alura.bytebank.teste
+
+import br.com.alura.bytebank.modelo.Endereco
+
+fun testaAny() {
+    val endereco = Endereco(
+        logradouro = "Rua Vergueiro",
+        complemento = "Alura",
+        cep = "00000-000"
+    )
+    val enderecoNovo = Endereco(
+        logradouro = "Rua Vergueiro",
+        complemento = "Alura",
+        cep = "00000-000"
+    )
+
+    println(endereco.equals(enderecoNovo))
+
+    println(endereco.hashCode())
+    println(enderecoNovo.hashCode())
+
+    println(endereco)
+    println(enderecoNovo)
+
+    println(
+        "${endereco.javaClass}@${
+            Integer.toHexString(endereco.hashCode())
+        }"
+    )
+}
